+++ src/components/RadioGroup/RadioGroup.module.scss
@@ -55,3 +55,10 @@ label {
         font-size: 0.9rem;
     }
 }
+
+.displayNone{
+    display: none;
+}
+.displayBlock{
+    display: block;
+}
