+++ src/components/helpers/handleKeyPress.js
@@ -1,0 +1,7 @@
+const handleKeyPress = (event) => {
+    if (event.key === 'Enter' && event.target.value !== '') {
+        event.preventDefault();
+    }
+}
+
+export default handleKeyPress;
